Status Upload
Status folder, and servers.php are all for BEC - Status is a plugin for
Battleeye Extended Controls, and it displays the status of the arma 3
server on a web page. servers.php is a modify version of the wipcode.php
for the use of e107 - will need to be modify.